Report error when server fails to stop, waits server to stop. (#203)
* Simplify autoload code of bootstrap.php.

* Remove ineffective use statement.

* Wait for the specified number of seconds when the server fails to stop.

* Report error when server fails to stop.

* Update SignalDispatcher.

* Update StopCommand.php

+++ src/Commands/StopCommand.php
@@ -58,7 +58,11 @@ class StopCommand extends Command
 
         $this->info('Stopping server...');
 
-        $inspector->stopServer();
+        if (! $inspector->stopServer()) {
+            $this->error('Failed to stop Swoole server.');
+
+            return 1;
+        }
 
         app(SwooleServerStateFile::class)->delete();
